requirements.txt
blinker==1.9.0
click==8.2.0
dotenv==0.9.9
Flask==3.1.1
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
